requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
sentence-transformers
pypdf
tiktoken
aiofiles
fastapi
uvicorn
jinja2
python-multipart
PyPDF2
faiss-cpu
python-dotenv
psutil